+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.0.0
 
+* Configure a serial console with attribute `node.serial`.
 * Deploy cgroups with attributes in `node.sys.cgroups`.
 * Alter Grub boot configuration attributes in `node.sys.boot`. 
 * System login banner with attributes in `node.sys.banner`,